[layout] Use NEON SIMD instructions for apply_forward()
To skip four glyphs at a time.

Unfortunately slows down benchmark-shape by up to 25% for Amiri, instead of
speeding anything up. This is due to the overhead of switching back and forth
between the ARM and NEON processors.

To be experimented with: Same architecture but with AVX on x86_64 to skip 8
glyphs at a time.

https://github.com/harfbuzz/harfbuzz/issues/566